refactoring: 신고 기능 통합으로 인한 프로젝트 정리
