--- matzipback/src/main/java/com/matzip/matzipback/report/command/domain/repository/ListDomainRepository.java
@@ -1,12 +1,0 @@
-package com.matzip.matzipback.report.command.domain.repository;
-
-import com.matzip.matzipback.report.command.domain.aggregate.Report;
-
-public interface ListDomainRepository {
-
-    boolean existsByReporterUserSeqAndListSeq(Long reporterUserSeq, Long listSeq);
-
-    Report save(Report newReport);
-
-    boolean existsByReporterUserSeqAndListCommentSeq(Long reporterUserSeq, Long listCommentSeq);
-}
